refactor: extract pages into app/pages/, delete empty app/blueprints/
Tenth and final phase of the courses.py/main.py decomposition. What's
left of main.py after Phases 6-9 carved out kbm/discussion/content/
courses/quiz/assignment is the true generic-pages remnant: index,
dashboard, settings, history, privacy-policy, sponsor, and the
set-language API — moved as-is to app/pages/routes.py.

Deliberately kept the Blueprint's registered name as 'main' (not
renamed to 'pages') even though the file moved: these 6 routes are
linked to from nearly every template in the app (base.html, _sidebar.html,
settings.html, and the templates of every single feature app), and
Flask blueprint names are independent of where the module physically
lives. Renaming would have meant updating ~20 more `url_for('main.dashboard')`
/ `url_for('main.settings')` style references for zero functional
benefit — Django itself doesn't care what internal Python identifier
a view module uses either, only the registered route/endpoint matters.
Templates for these routes (index.html, settings.html, history.html,
privacy_policy.html, sponsor.html) stay in the shared app/templates/
root too, for the same reason: index.html in particular is also
rendered directly by app/auth/routes.py, so it's genuinely a shared
page, not pages-app-exclusive.

app/blueprints/ (empty since Phase 5, holding only an emptied
__init__.py) is deleted now that main.py was its last real file.
app/core/registry.py's docstring updated to reflect that the per-app
restructure is complete.

Verified: create_app() builds with all 222 routes intact (pure file
move, zero route changes), SQLAlchemy metadata still has 33 tables,
pytest unchanged from baseline (25 pre-existing failures, 99 passing),
and a scripted smoke test hits every remaining page route (index,
dashboard, settings, history, privacy-policy, sponsor, set-language)
plus a login flow — all resolve correctly.

This completes the full Django-style per-app restructure: app/models/
and app/blueprints/ (the original two "everything lives here" folders)
are gone, replaced by 20 self-contained app/<feature>/ packages
(models.py + routes.py + services.py where applicable + templates/<feature>/
+ tests/), a shared app/core/ project layer, and a handful of
deliberately-centralized things (translations, static assets, the
0-model cross-cutting tables, email_service.py) with documented reasons
for staying put.
